+++ akeneo-design-system/src/components/Input/TextAreaInput/TextAreaInput.unit.tsx
@@ -3,16 +3,9 @@ import {TextAreaInput} from './TextAreaInput';
 import {fireEvent, render, screen} from '../../../storybook/test-util';
 import {ContentBlock} from 'draft-js';
 
-jest.mock('draft-js', () => {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-  const draftJS = jest.requireActual('draft-js');
-
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-  return {
-    ...draftJS,
-    convertFromHTML: (text: string) => ('WILL FAIL' === text ? undefined : {contentBlocks: [new ContentBlock({text})]}),
-  };
-});
+jest.mock('html-to-draftjs', () => (text: string) =>
+  'WILL FAIL' === text ? undefined : {contentBlocks: [new ContentBlock({text})]}
+);
 
 test('it renders and handle changes', () => {
   const handleChange = jest.fn();
@@ -65,7 +58,7 @@ test('it renders and displays the character left label', () => {
   expect(screen.getByText('100 character remaining')).toBeInTheDocument();
 });
 
-test('it renders a rich text editor', () => {
+test('it renders rich text editor and handle changes', () => {
   const handleChange = jest.fn();
 
   render(
@@ -77,6 +70,19 @@ test('it renders a rich text editor', () => {
 
   expect(screen.getByLabelText('rdw-wrapper')).toBeInTheDocument();
   expect(screen.getByText('Nice RTF content')).toBeInTheDocument();
+  expect(screen.getByRole('textbox')).toBeInTheDocument();
+
+  const input = screen.getByRole('textbox') as HTMLInputElement;
+
+  // Hack to trigger onBeforeInput event listened by DraftJS
+  fireEvent.paste(input, {
+    clipboardData: {
+      types: ['text/plain'],
+      getData: () => 'New ',
+    },
+  });
+
+  expect(handleChange).toHaveBeenCalledWith('<p>New Nice RTF content</p>\n');
 });
 
 test('it renders an empty rich text editor when the value is unprocessable', () => {