+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 - PIM-9718: Decimals attribute values with no separators are well formatted
 - PIM-9727: Add missing query params to hatoas links
 - API-9698: Refresh ES index after creating a product from the UI in order to well send product created event to event subscriptions
+- PIM-9711: Check that a category root isn't linked to a user or a channel before moving it to a sub-category
+- PIM-9730: Fix category tree initialization in the PEF when switching tabs
+- PIM-9679: Clean existing text attribute values removing linebreaks
 
 ## New features